+++ applications/lethe-fluid-particles/cfd_dem_coupling.cc
@@ -16,6 +16,8 @@
 
 #include "fem-dem/cfd_dem_coupling.h"
 
+#include <core/parameters.h>
+
 int
 main(int argc, char *argv[])
 {
@@ -31,13 +33,15 @@ main(int argc, char *argv[])
           std::exit(1);
         }
 
-      const unsigned int dim = get_dimension(argv[1]);
+      const unsigned int                  dim = get_dimension(argv[1]);
+      const Parameters::SizeOfSubsections size_of_subsections =
+        Parameters::get_size_of_subsections(argv[1]);
 
       if (dim == 2)
         {
           ParameterHandler              prm;
           CFDDEMSimulationParameters<2> NSparam;
-          NSparam.declare(prm);
+          NSparam.declare(prm, size_of_subsections);
           // Parsing of the file
           prm.parse_input(argv[1]);
           NSparam.parse(prm);
@@ -50,7 +54,7 @@ main(int argc, char *argv[])
         {
           ParameterHandler              prm;
           CFDDEMSimulationParameters<3> NSparam;
-          NSparam.declare(prm);
+          NSparam.declare(prm, size_of_subsections);
           // Parsing of the file
           prm.parse_input(argv[1]);
           NSparam.parse(prm);